+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/serialization/StandardSerializers.java
@@ -23,6 +23,8 @@ import com.exonum.binding.common.message.TransactionMessage;
 import com.google.protobuf.CodedOutputStream;
 import com.google.protobuf.MessageLite;
 
+import java.time.ZonedDateTime;
+
 /**
  * A collection of pre-defined serializers.
  */
@@ -145,6 +147,13 @@ public final class StandardSerializers {
    */
   public static Serializer<TransactionMessage> transactionMessage() {
     return TransactionMessageSerializer.INSTANCE;
+  }
+
+  /**
+   * Returns a serializer of ZonedDateTime instances.
+   */
+  public static Serializer<ZonedDateTime> zonedDateTime() {
+    return ZonedDateTimeSerializer.INSTANCE;
   }
 
   /**