+++ exonum-java-binding-core/rust/ejb-app/src/main.rs
@@ -1,5 +1,6 @@
 extern crate exonum_btc_anchoring;
 extern crate exonum_configuration;
+extern crate exonum_time;
 extern crate java_bindings;
 extern crate serde;
 #[macro_use]